+++ modules/canbus/vehicle/lincoln/protocol/steering_65.cc
@@ -71,7 +71,8 @@ void Steering65::Parse(const std::uint8_t *bytes, int32_t length,
                        const struct timeval &timestamp,
                        ChassisDetail *chassis_detail) const {
   chassis_detail->mutable_eps()->set_timestamp_65(
-      timestamp.tv_sec + timestamp.tv_usec / 1000000.0);
+      static_cast<double>(timestamp.tv_sec) +
+      static_cast<double>(timestamp.tv_usec) / 1000000.0);
   Parse(bytes, length, chassis_detail);
 }
